Extract the hook docblock processing so it can be reused.
