+++ src/ApplyFiltersDynamicFunctionReturnTypeExtension.php
@@ -11,18 +11,17 @@ namespace SzepeViktor\PHPStan\WordPress;
 use PhpParser\Node\Expr\FuncCall;
 use PHPStan\Analyser\Scope;
 use PHPStan\Reflection\FunctionReflection;
-use PHPStan\Type\FileTypeMapper;
 use PHPStan\Type\Type;
 use PHPStan\Type\MixedType;
 
 class ApplyFiltersDynamicFunctionReturnTypeExtension implements \PHPStan\Type\DynamicFunctionReturnTypeExtension
 {
-    /** @var \PHPStan\Type\FileTypeMapper */
-    protected $fileTypeMapper;
+    /** @var HookDocBlock */
+    protected $hookDocBlock;
 
-    public function __construct(FileTypeMapper $fileTypeMapper)
+    public function __construct(HookDocBlock $hookDocBlock)
     {
-        $this->fileTypeMapper = $fileTypeMapper;
+        $this->hookDocBlock = $hookDocBlock;
     }
 
     public function isFunctionSupported(FunctionReflection $functionReflection): bool
@@ -42,25 +41,11 @@ class ApplyFiltersDynamicFunctionReturnTypeExtension implements \PHPStan\Type\Dy
     public function getTypeFromFunctionCall(FunctionReflection $functionReflection, FuncCall $functionCall, Scope $scope): Type
     {
         $default = new MixedType();
-        $comment = self::getNullableNodeComment($functionCall);
+        $resolvedPhpDoc = $this->hookDocBlock->getNullableHookDocBlock($functionCall, $scope);
 
-        if ($comment === null) {
+        if ($resolvedPhpDoc === null) {
             return $default;
         }
-
-        // Fetch the docblock contents.
-        $code = $comment->getText();
-
-        // Resolve the docblock in scope.
-        $classReflection = $scope->getClassReflection();
-        $traitReflection = $scope->getTraitReflection();
-        $resolvedPhpDoc = $this->fileTypeMapper->getResolvedPhpDoc(
-            $scope->getFile(),
-            ($scope->isInClass() && $classReflection !== null) ? $classReflection->getName() : null,
-            ($scope->isInTrait() && $traitReflection !== null) ? $traitReflection->getName() : null,
-            $scope->getFunctionName(),
-            $code
-        );
 
         // Fetch the `@param` values from the docblock.
         $params = $resolvedPhpDoc->getParamTags();
@@ -70,24 +55,5 @@ class ApplyFiltersDynamicFunctionReturnTypeExtension implements \PHPStan\Type\Dy
         }
 
         return $default;
-    }
-
-    private static function getNullableNodeComment(FuncCall $node): ?\PhpParser\Comment\Doc
-    {
-        $startLine = $node->getStartLine();
-
-        while ($node !== null && $node->getStartLine() === $startLine) {
-            // Fetch the docblock from the node.
-            $comment = $node->getDocComment();
-
-            if ($comment !== null) {
-                return $comment;
-            }
-
-            /** @var \PhpParser\Node|null */
-            $node = $node->getAttribute('parent');
-        }
-
-        return null;
     }
 }
